Differentiate node vs webapi metadata
