+++ packages/atxp-server/src/node/index.ts
@@ -2,5 +2,5 @@
 // These functions adapt Node.js HTTP APIs to use the core business logic
 
 export { checkToken } from './token.js';
-export { sendOAuthChallenge } from './oauth.js';
+export { sendOAuthChallenge, sendProtectedResourceMetadata, sendOAuthMetadata } from './oauth.js';
 export { parseMcpRequests, parseBody } from './http.js';
