+++ packages/atxp-server/src/webapi/oauth.ts
@@ -1,5 +1,6 @@
-import { TokenCheck } from "../types.js";
+import { ProtectedResourceMetadata, TokenCheck } from "../types.js";
 import { createOAuthChallengeResponseCore } from "../core/oauth.js";
+import * as oauth from 'oauth4webapi';
 
 /**
  * Web API implementation of OAuth challenge sending for Cloudflare Workers, Deno, etc.
@@ -17,5 +18,31 @@ export function sendOAuthChallengeWebApi(tokenCheck: TokenCheck): Response | nul
   return new Response(responseData.body, {
     status: responseData.status,
     headers: responseData.headers
+  });
+}
+
+export function sendProtectedResourceMetadata(metadata: ProtectedResourceMetadata | null): Response | null {
+  if (!metadata) {
+    return null;
+  }
+
+  return new Response(JSON.stringify(metadata), {
+    status: 200,
+    headers: {
+      'Content-Type': 'application/json'
+    }
+  });
+}
+
+export function sendOAuthMetadata(metadata: oauth.AuthorizationServer | null): Response | null {
+  if (!metadata) {
+    return null;
+  }
+  
+  return new Response(JSON.stringify(metadata), {
+    status: 200,
+    headers: {
+      'Content-Type': 'application/json'
+    }
   });
 }
